Replace the symbolic link of jrlMathTools by real files.
The symbolic link does not work under Win32
As a result, the files created just contained the path
of the file linked, e.g. ../jrl/mathtools/vector3.h

To correct this issue, the files are replaced by a direct
include of the corresponding file
e.g. #include "../jrl/mathtools/vector3.h"

These headers are deprecated and will be removed in the future.